Move knx base entity to separate module
