+++ homeassistant/components/knx/binary_sensor.py
@@ -24,7 +24,7 @@ from homeassistant.helpers.typing import ConfigType
 
 from . import KNXModule
 from .const import ATTR_COUNTER, ATTR_SOURCE, KNX_MODULE_KEY
-from .knx_entity import KnxYamlEntity
+from .entity import KnxYamlEntity
 from .schema import BinarySensorSchema
