+++ homeassistant/components/knx/cover.py
@@ -27,7 +27,7 @@ from homeassistant.helpers.typing import ConfigType
 
 from . import KNXModule
 from .const import KNX_MODULE_KEY
-from .knx_entity import KnxYamlEntity
+from .entity import KnxYamlEntity
 from .schema import CoverSchema
