+++ homeassistant/components/knx/entity.py
