+++ homeassistant/components/knx/select.py
@@ -29,7 +29,7 @@ from .const import (
     KNX_ADDRESS,
     KNX_MODULE_KEY,
 )
-from .knx_entity import KnxYamlEntity
+from .entity import KnxYamlEntity
 from .schema import SelectSchema
